Thu, Dec 21, 2023, 4:22 PM -05:00

+++ half/half.c
@@ -16,8 +16,8 @@ int main(void)
 }
 
 // TODO: Complete the function
-    double half(float bill, float tax, int tip)
-{   double myHalf = (bill + (bill * tax * .01) + (bill * tip * .01)) / 2.0;
+    float half(float bill, float tax, int tip)
+{   float myHalf = (bill + (bill * tax * .01) + (bill * tip * .01)) / 2.0;
     return myHalf;
 
 }
